update vendor group decorator

+++ trips/views.py
@@ -38,55 +38,64 @@ def each_trip(request, trip_id):
 @login_required
 # @group_required('vendor')
 def read_create_trip(request):
-    trips = Trip.objects.all()
-    if request.method == "POST":
+    if request.user.groups.filter(name='vendor').exists():
+        trips = Trip.objects.all()
+        if request.method == "POST":
 
-        create_form = TripForm(request.POST)
-        if create_form.is_valid():
-            create_form.save()
-            return redirect(reverse(read_create_trip))
+            create_form = TripForm(request.POST)
+            if create_form.is_valid():
+                create_form.save()
+                return redirect(reverse(read_create_trip))
+            else:
+                return render(request, 'trips/index.template.html', {
+                    "form": create_form,
+                    'trips':trips
+                })
         else:
+            create_form = TripForm()
             return render(request, 'trips/index.template.html', {
                 "form": create_form,
                 'trips':trips
             })
     else:
-        create_form = TripForm()
-        return render(request, 'trips/index.template.html', {
-            "form": create_form,
-            'trips':trips
-        })
+        return redirect(reverse('home_route'))
 
 @login_required
 # @group_required('vendor')
 def update_trip(request, trip_id):
-    updating_trip = get_object_or_404(Trip, pk=trip_id)
-    if request.method == "POST":
-        update_form = TripForm(request.POST, instance=updating_trip)
-        if update_form.is_valid():
-            update_form.save()
-            return redirect(reverse(read_create_trip))
+    if request.user.groups.filter(name='vendor').exists():
+        updating_trip = get_object_or_404(Trip, pk=trip_id)
+        if request.method == "POST":
+            update_form = TripForm(request.POST, instance=updating_trip)
+            if update_form.is_valid():
+                update_form.save()
+                return redirect(reverse(read_create_trip))
+            else:
+                return render(request, 'trips/update.template.html', {
+                    "form": update_form
+                })
         else:
+            update_form = TripForm(instance=updating_trip)
             return render(request, 'trips/update.template.html', {
                 "form": update_form
             })
     else:
-        update_form = TripForm(instance=updating_trip)
-        return render(request, 'trips/update.template.html', {
-            "form": update_form
-        })
+        return redirect(reverse('home_route'))
 
 @login_required
 # @group_required('vendor')
 def delete_trip(request, trip_id):
-    trip_to_delete = get_object_or_404(Trip, pk=trip_id)
-    if request.method == 'POST':
-        trip_to_delete.delete()
-        return redirect(read_create_trip)
+    if request.user.groups.filter(name='vendor').exists():
+        trip_to_delete = get_object_or_404(Trip, pk=trip_id)
+        if request.method == 'POST':
+            trip_to_delete.delete()
+            return redirect(read_create_trip)
+        else:
+            return render(request, 'trips/delete.template.html', {
+                "trip": trip_to_delete
+            })
     else:
-        return render(request, 'trips/delete.template.html', {
-            "trip": trip_to_delete
-        })
+        return redirect(reverse('home_route'))
 
 
 def search(request):
